+++ src/stack/repos/fetch_stack.py
@@ -22,9 +22,9 @@ import click
 
 from git import exc
 
-from stack.build.build_util import host_and_path_for_repo
 from stack.config.util import get_config_setting, get_dev_root_path, verbose_enabled
 from stack.repos.setup_repositories import process_repo
+from stack.repos.repo_util import host_and_path_for_repo
 from stack.util import error_exit
